refactor(api): dedicated email controller
